requirements.txt
streamlit
supabase
python-dotenv
google-generativeai
requests
Pillow
